fix bug on delete bug opening multiple modals
