+++ src/components/GoalList.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Goal from "./Goal";
+
+function DisplayGoals({ goals, setGoals }) {
+  const renderGoals = () => {
+    return goals.map((goal) => <Goal goal={goal} setGoals={setGoals} />);
+  };
+
+  return <div>{renderGoals()}</div>;
+}
+
+export default DisplayGoals;
